adds the setSelectedGeographyVariableStore functionality
